IMP - Refactor analysis edit common handler
